rename `secret.rs` to `protected.rs`
