+++ src/global/protected.rs
